Resolve compiler warnings related to signed/unsigned comparsions

+++ src/game.cpp
@@ -375,10 +375,10 @@ void Render() {
               int bufId = ScrWork[SW_CHA1SURF + i];
               Characters2D[bufId].Render(i, layer);
             }
-            if (ScrWork[6361] == layer && ScrWork[6360]) {
+            if (ScrWork[6361] == static_cast<int>(layer) && ScrWork[6360]) {
               UI::MapSystem::Render();
             }
-            if (ScrWork[SW_MASK1PRI] == layer) {
+            if (ScrWork[SW_MASK1PRI] == static_cast<int>(layer)) {
               int maskAlpha =
                   ScrWork[SW_MASK1ALPHA_OFS] + ScrWork[SW_MASK1ALPHA];
               if (maskAlpha) {
@@ -400,7 +400,8 @@ void Render() {
             }
 
             if (Profile::UseScreenCapEffects) {
-              if (ScrWork[SW_EFF_CAP_BUF] && ScrWork[SW_EFF_CAP_PRI] == layer) {
+              if (ScrWork[SW_EFF_CAP_BUF] &&
+                  ScrWork[SW_EFF_CAP_PRI] == static_cast<int>(layer)) {
                 int bufId = (int)std::log2(ScrWork[SW_EFF_CAP_BUF]);
                 if (Backgrounds2D[bufId]->Status == LS_Loaded) {
                   Renderer->CaptureScreencap(Backgrounds2D[bufId]->BgSprite);
@@ -408,7 +409,7 @@ void Render() {
               }
 
               if (ScrWork[SW_EFF_CAP_BUF2] &&
-                  ScrWork[SW_EFF_CAP_PRI2] == layer) {
+                  ScrWork[SW_EFF_CAP_PRI2] == static_cast<int>(layer)) {
                 int bufId = (int)std::log2(ScrWork[SW_EFF_CAP_BUF2]);
                 if (Backgrounds2D[bufId]->Status == LS_Loaded) {
                   Renderer->CaptureScreencap(Backgrounds2D[bufId]->BgSprite);
@@ -418,7 +419,7 @@ void Render() {
 
             if (Profile::UseMoviePriority &&
                 (Profile::GameFeatures & GameFeature::Video) &&
-                ScrWork[SW_MOVIEPRI] == layer) {
+                ScrWork[SW_MOVIEPRI] == static_cast<int>(layer)) {
               Video::VideoRender(ScrWork[SW_MOVIEALPHA] / 256.0f);
             }
           }
