+++ src/renderer/vulkan/3d/renderable3d.cpp
@@ -324,7 +324,7 @@ void Renderable3D::MakePlane() {
 
 void Renderable3D::InitMeshAnimStatus() {
   int totalMorphedVertices = 0;
-  for (int i = 0; i < StaticModel->MeshCount; i++) {
+  for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
     MeshAnimStatus[i].MorphedVerticesOffset = totalMorphedVertices;
     if (StaticModel->Meshes[i].MorphTargetCount > 0) {
       totalMorphedVertices += StaticModel->Meshes[i].VertexCount;
@@ -361,7 +361,7 @@ void Renderable3D::ReloadDefaultMeshAnimStatus() {
   VertexBufferDaSH* currentMorphedVertexDaSH =
       (VertexBufferDaSH*)currentMorphedVertex;
 
-  for (int i = 0; i < StaticModel->MeshCount; i++) {
+  for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
     MeshAnimStatus[i].Visible = 1.0f;
     if (StaticModel->Meshes[i].MorphTargetCount > 0) {
       for (int j = 0; j < StaticModel->Meshes[i].MorphTargetCount; j++) {
@@ -374,7 +374,7 @@ void Renderable3D::ReloadDefaultMeshAnimStatus() {
 void Renderable3D::SwitchAnimation(int16_t animId, float transitionTime) {
   if (Animator.CurrentAnimation != 0 && transitionTime > 0.0f) {
     PrevPoseWeight = 1.0f;
-    for (int i = 0; i < StaticModel->BoneCount; i++) {
+    for (uint32_t i = 0; i < StaticModel->BoneCount; i++) {
       PrevBoneTransforms[i] = CurrentPose[i].LocalTransform;
     }
     memcpy(PrevMeshAnimStatus, MeshAnimStatus, sizeof(MeshAnimStatus));
@@ -386,7 +386,7 @@ void Renderable3D::SwitchAnimation(int16_t animId, float transitionTime) {
 }
 
 void Renderable3D::ReloadDefaultBoneTransforms() {
-  for (int i = 0; i < StaticModel->BoneCount; i++) {
+  for (uint32_t i = 0; i < StaticModel->BoneCount; i++) {
     CurrentPose[i].LocalTransform = StaticModel->Bones[i].BaseTransform;
   }
 }
@@ -423,7 +423,7 @@ void Renderable3D::CalculateMorphedVertices(int id) {
   VertexBuffer* currentVertexRNE = (VertexBuffer*)currentVertex;
   VertexBufferDaSH* currentVertexDaSH = (VertexBufferDaSH*)currentVertex;
 
-  for (int j = 0; j < mesh->VertexCount; j++) {
+  for (uint32_t j = 0; j < mesh->VertexCount; j++) {
     if (Profile::Scene3D::Version == +LKMVersion::DaSH) {
       memcpy(currentMorphedVertexDaSH, currentVertexDaSH,
              sizeof(VertexBufferDaSH));
@@ -476,7 +476,7 @@ void Renderable3D::CalculateMorphedVertices(int id) {
         StaticModel->MorphVertexBuffers +
         StaticModel->MorphTargets[mesh->MorphTargetIds[k]].VertexOffset;
 
-    for (int j = 0; j < mesh->VertexCount; j++) {
+    for (uint32_t j = 0; j < mesh->VertexCount; j++) {
       if (Profile::Scene3D::Version == +LKMVersion::DaSH) {
         currentMorphedVertexDaSH->Position +=
             (currentMorphTargetVbo->Position - currentVertexDaSH->Position) *
@@ -554,7 +554,7 @@ void Renderable3D::Update(float dt) {
     if (PrevPoseWeight < 0.0f) PrevPoseWeight = 0.0f;
   }
 
-  for (int i = 0; i < StaticModel->MeshCount; i++) {
+  for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
     CalculateMorphedVertices(i);
   }
 }
@@ -834,7 +834,7 @@ void Renderable3D::Render() {
   CurrentWriteDescriptorSet = 0;
 
   for (int i = RP_First; i < RP_Count; i++) {
-    for (int j = 0; j < StaticModel->MeshCount; j++) {
+    for (size_t j = 0; j < StaticModel->MeshCount; j++) {
       DrawMesh(j, (RenderPass)i);
     }
   }
@@ -934,7 +934,7 @@ void Renderable3D::LoadMeshUniforms(int id) {
     if (!UniformsUpdated[id]) {
       if (mesh.UsedBones > 0) {
         glm::mat4* outBone = entry.Bones;
-        for (int j = 0; j < mesh.UsedBones; j++) {
+        for (uint32_t j = 0; j < mesh.UsedBones; j++) {
           memcpy(outBone, glm::value_ptr(CurrentPose[mesh.BoneMap[j]].Offset),
                  sizeof(glm::mat4));
           outBone++;
@@ -967,16 +967,16 @@ void Renderable3D::UnloadSync() {
     ImpLog(LL_Info, LC_Renderable3D, "Unloading model %d\n", StaticModel->Id);
     if (IsSubmitted) {
       if (StaticModel->Type == ModelType_Background) {
-        for (int i = 0; i < StaticModel->MeshCount; i++) {
-          for (int j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
+        for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
+          for (uint32_t j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
             vmaUnmapMemory(Allocator, BackgroundMvpBuffers[i][j].Allocation);
             vmaDestroyBuffer(Allocator, BackgroundMvpBuffers[i][j].Buffer,
                              BackgroundMvpBuffers[i][j].Allocation);
           }
         }
       } else {
-        for (int i = 0; i < StaticModel->MeshCount; i++) {
-          for (int j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
+        for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
+          for (uint32_t j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
             vmaUnmapMemory(Allocator, MeshUniformBuffers[i][j].Allocation);
             vmaDestroyBuffer(Allocator, MeshUniformBuffers[i][j].Buffer,
                              MeshUniformBuffers[i][j].Allocation);
@@ -1018,8 +1018,8 @@ void Renderable3D::MainThreadOnLoad() {
          StaticModel->Id);
 
   if (StaticModel->Type == ModelType_Background) {
-    for (int i = 0; i < StaticModel->MeshCount; i++) {
-      for (int j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
+    for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
+      for (uint32_t j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
         BackgroundMvpBuffers[i][j] = CreateBuffer(
             sizeof(BgMVPUniformBufferType), VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT,
             VMA_MEMORY_USAGE_CPU_ONLY);
@@ -1046,8 +1046,8 @@ void Renderable3D::MainThreadOnLoad() {
   void* stagingBufferMapped;
   vmaMapMemory(Allocator, stagingBuffer.Allocation, &stagingBufferMapped);
 
-  for (int i = 0; i < StaticModel->MeshCount; i++) {
-    for (int j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
+  for (uint32_t i = 0; i < StaticModel->MeshCount; i++) {
+    for (uint32_t j = 0; j < MAX_FRAMES_IN_FLIGHT; j++) {
       MeshUniformBuffers[i][j] = CreateBuffer(
           sizeof(MeshUniformBufferType), VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT,
           VMA_MEMORY_USAGE_CPU_ONLY);
@@ -1126,7 +1126,7 @@ void Renderable3D::MainThreadOnLoad() {
     }
   }
 
-  for (int i = 0; i < StaticModel->TextureCount; i++) {
+  for (uint32_t i = 0; i < StaticModel->TextureCount; i++) {
     TexBuffers[i] = StaticModel->Textures[i].Submit();
     if (TexBuffers[i] == 0) {
       ImpLog(LL_Fatal, LC_Renderable3D,
