+++ src/video/ffmpegplayer.cpp
@@ -24,7 +24,8 @@ int StreamRead(void* ptr, uint8_t* buf, int buf_size) {
   Io::InputStream* stream = reinterpret_cast<Io::InputStream*>(ptr);
 
   uint64_t bytesRead = stream->Read(buf, buf_size);
-  if ((bytesRead == IoError_Fail || bytesRead == IoError_Eof))
+  if ((bytesRead == static_cast<uint64_t>(IoError_Fail) ||
+       bytesRead == static_cast<uint64_t>(IoError_Eof)))
     return AVERROR_EOF;
 
   return bytesRead;
@@ -161,7 +162,7 @@ void FFmpegPlayer::Play(Io::InputStream* stream, bool looping, bool alpha) {
   if (audioStreamId != AVERROR_STREAM_NOT_FOUND)
     audioStream = FormatContext->streams[audioStreamId];
 
-  for (int j = 0; j < FormatContext->nb_streams; ++j) {
+  for (int j = 0; j < static_cast<int>(FormatContext->nb_streams); ++j) {
     FormatContext->streams[j]->discard =
         j == videoStreamId || j == audioStreamId ? AVDISCARD_DEFAULT
                                                  : AVDISCARD_ALL;
